+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Coffee Machine Library Changelog
 
+## []
+### Removed
+- Removed unused configuration parameters
+
 ## [0.3.0] - 24/03/2021
 ### Changed
 - Serial interface output can now be configured to be on/off